iq-cli-derive: Factor Options-related code into options module
